requirements.txt
# GitTalker Production Dependencies
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# AI/ML
openai==1.3.9
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.4

# Communication
slack-sdk==3.26.1

# HTTP & Utilities
httpx==0.25.2
python-dotenv==1.0.0

# Development & Quality (optional)
ruff==0.1.6
mypy==1.7.1